Add constructor in Beacon
