+++ packages/core-contracts/test/contracts/proxy/BeaconProxy.test.js
@@ -6,10 +6,16 @@ const { findEvent } = require('@synthetixio/core-js/utils/events');
 describe('BeaconProxy', () => {
   let BeaconProxy, Instance, Beacon, Implementation;
 
+  let owner;
+
+  before('identify signers', async () => {
+    [owner] = await ethers.getSigners();
+  });
+
   describe('when deploying the proxy and setting implementation A as the first implementation', () => {
     before('deploy the beacon', async () => {
-      const factory = await ethers.getContractFactory('BeaconMock');
-      Beacon = await factory.deploy();
+      const factory = await ethers.getContractFactory('Beacon');
+      Beacon = await factory.deploy(owner.address);
     });
 
     before('deploy the implementation', async () => {
@@ -18,15 +24,13 @@ describe('BeaconProxy', () => {
     });
 
     before('set the beacons first implementation', async () => {
-      const tx = await Beacon.setImplementation(Implementation.address);
+      const tx = await Beacon.upgradeTo(Implementation.address);
       await tx.wait();
     });
 
     before('deploy the proxy and set the beacon', async () => {
-      const factory = await ethers.getContractFactory('BeaconProxyMock');
-      BeaconProxy = await factory.deploy();
-      const tx = await BeaconProxy.setBeacon(Beacon.address);
-      await tx.wait();
+      const factory = await ethers.getContractFactory('BeaconProxy');
+      BeaconProxy = await factory.deploy(Beacon.address);
 
       Instance = await ethers.getContractAt('ImplementationMockA', BeaconProxy.address);
     });
@@ -57,49 +61,21 @@ describe('BeaconProxy', () => {
       });
     });
 
-    describe('when trying to upgrade to an EOA', () => {
-      it('reverts', async () => {
-        const [user] = await ethers.getSigners();
-        await assertRevert(
-          Beacon.setImplementation(user.address),
-          `InvalidContract("${user.address}")`
-        );
-      });
-    });
-
-    describe('when trying to upgrade to the zero address', () => {
-      it('reverts', async () => {
-        const zeroAddress = '0x0000000000000000000000000000000000000000';
-        await assertRevert(
-          Beacon.setImplementation(zeroAddress),
-          `InvalidAddress("${zeroAddress}")`
-        );
-      });
-    });
-
     describe('when upgrading to implementation B', () => {
-      let upgradeReceipt;
-
       before('deploy the implementation', async () => {
         const factory = await ethers.getContractFactory('ImplementationMockB');
         Implementation = await factory.deploy();
       });
 
       before('upgrade', async () => {
-        const tx = await Beacon.setImplementation(Implementation.address);
-        upgradeReceipt = await tx.wait();
+        const tx = await Beacon.upgradeTo(Implementation.address);
+        await tx.wait();
 
         Instance = await ethers.getContractAt('ImplementationMockB', BeaconProxy.address);
       });
 
       it('shows that the implementation is set', async () => {
         assert.equal(await Beacon.getImplementation(), Implementation.address);
-      });
-
-      it('emitted an Upgraded event', async () => {
-        const event = findEvent({ receipt: upgradeReceipt, eventName: 'Upgraded' });
-
-        assert.equal(event.args.implementation, Implementation.address);
       });
 
       describe('when interacting with implementation B', () => {
